refactor(coordinator): extract request introspection helpers from consumer.go
Move the pure request-introspection / token-estimation cluster out of the 5.7k-LOC
consumer.go into a focused request_introspection.go (token + billing cost
estimation, media/tool detection, cache-affinity keys, provider-serial allowlist
parsing, and the media-cost constants). Same package, no behavior change — these
functions have no Server dependency. consumer.go drops from ~5700 to ~5311 LOC.
